Enrich the types for protocol data
We now distinguish protocol data read from an XML file, the protocol
data we need at runtime, and protocol data we need at runtime knotted
together for easier processing.

The knotted variant has the <arg>s refer to actual Interface and ArgEnum
objects, rather than by string names. This means that when we want to
track which freshly created wayland object ID has which interface, we
can simply read this from a value associated with the creating message.

Using this, we can start to write a Haskell API